Added ProductService

+++ TechnoMarket/src/main/java/com/technomarket/model/services/ProductService.java
@@ -25,7 +25,7 @@ public class ProductService {
         }
     }
 
-    public ProductWithAllDTO getProduct(long productId) {
+    public ProductWithAllDTO getProduct(int productId) {
 
         //TODO
         return null;
